requirements.txt
numpy
pandas
matplotlib
nltk
random
pickle
streamlit
pip
scikit-learn
sklearn
pillow